requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
reportlab==4.0.7
pillow==10.1.0
pydantic==2.5.2
